requirements.txt
numpy
opencv-python
requests
Pillow
tqdm